+++ src/react/hooks/useStore.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+import Context from "../components/Context";
+import Store from "../../interfaces/Store";
+
+export function useStore(): Store {
+  return useContext(Context);
+}